Fix loss of numeric precision. Thank you very much for your help!

+++ src/main/java/org/scify/jedai/similarityjoins/characterbased/AllPairs.java
@@ -1,17 +1,17 @@
 /*
-* Copyright [2016-2020] [George Papadakis (gpapadis@yahoo.gr)]
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
+ * Copyright [2016-2020] [George Papadakis (gpapadis@yahoo.gr)]
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package org.scify.jedai.similarityjoins.characterbased;
 
@@ -94,7 +94,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
     public String getMethodName() {
         return "Character-based All Pairs";
     }
-    
+
     @Override
     public String getMethodParameters() {
         return getMethodName() + " involves two parameters:\n"
@@ -124,7 +124,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
         }
         return result;
     }
-    
+
     @Override
     public String getParameterDescription(int parameterId) {
         switch (parameterId) {
@@ -138,7 +138,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
                 return "invalid parameter id";
         }
     }
-    
+
     @Override
     public String getParameterName(int parameterId) {
         switch (parameterId) {
@@ -150,7 +150,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
                 return "invalid parameter id";
         }
     }
-    
+
     private void getTokens(int rangeBound) {
         final TIntIntMap freqMap = new TIntIntHashMap();
         for (int k = rangeBound; k < noOfEntities; k++) {
@@ -164,7 +164,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
         }
 
         final List<IntPair> packages = new ArrayList<>();
-        for (TIntIntIterator iterator = freqMap.iterator(); iterator.hasNext();) {
+        for (TIntIntIterator iterator = freqMap.iterator(); iterator.hasNext(); ) {
             iterator.advance();
             packages.add(new IntPair(iterator.key(), iterator.value()));
         }
@@ -236,7 +236,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
 
         return rangeBound;
     }
-    
+
     private List<Comparison> performJoin(int rangeBound) {
         final List<Comparison> executedComparisons = new ArrayList<>();
         for (int x = 0; x < rangeBound; x++) {
@@ -252,7 +252,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
                 int distance = getEditDistance(attributeValues.get(x), attributeValues.get(y), threshold);
                 if (distance <= threshold) {
                     final Comparison currentComp = getComparison(originalId[x], originalId[y]);
-                    currentComp.setUtilityMeasure(1 - distance / threshold);
+                    currentComp.setUtilityMeasure(1 - (double) distance / threshold);
                     executedComparisons.add(currentComp);
                 }
             }
@@ -263,7 +263,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
             int lastToken = -1;
 
             final TIntSet occurances = new TIntHashSet();
-            for (TIntIterator iter = tokens[k].iterator(); iter.hasNext();) {
+            for (TIntIterator iter = tokens[k].iterator(); iter.hasNext(); ) {
                 int token = iter.next();
                 if (count++ >= threshold * q + 1) {
                     break;
@@ -297,7 +297,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
                 index.put(token, item);
             }
 
-            for (TIntIterator setIterator = occurances.iterator(); setIterator.hasNext();) {
+            for (TIntIterator setIterator = occurances.iterator(); setIterator.hasNext(); ) {
                 int cand = setIterator.next();
                 if (isCleanCleanER) {
                     if (originalId[k] < datasetDelimiter && originalId[cand] < datasetDelimiter) { // both belong to dataset 1
@@ -366,7 +366,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
         }
         return executedComparisons;
     }
-    
+
     @Override
     public void setNextRandomConfiguration() {
         super.setNextRandomConfiguration();
@@ -377,7 +377,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
     public void setNumberedGridConfiguration(int iterationNumber) {
         int ngSizeIteration = iterationNumber / gridThreshold.getNumberOfConfigurations();
         q = (Integer) gridNGSize.getNumberedValue(ngSizeIteration);
-        
+
         int thrIteration = iterationNumber - ngSizeIteration * gridThreshold.getNumberOfConfigurations();
         threshold = (Integer) gridThreshold.getNumberedValue(thrIteration);
     }
